+++ main.tf
@@ -129,3 +129,15 @@ module "kms_key" {
 module "security_hub_master" {
   source = "./modules/security_hub"
 }
+
+resource "aws_iam_account_password_policy" "master" {
+  count                          = var.aws_create_account_password_policy ? 1 : 0
+  allow_users_to_change_password = var.aws_account_password_policy.allow_users_to_change
+  max_password_age               = var.aws_account_password_policy.max_age
+  minimum_password_length        = var.aws_account_password_policy.minimum_length
+  password_reuse_prevention      = var.aws_account_password_policy.reuse_prevention_history
+  require_lowercase_characters   = var.aws_account_password_policy.require_lowercase_characters
+  require_numbers                = var.aws_account_password_policy.require_numbers
+  require_symbols                = var.aws_account_password_policy.require_symbols
+  require_uppercase_characters   = var.aws_account_password_policy.require_uppercase_characters
+}